requirements.txt
fastmcp
pypandoc
python-dotenv
python-docx
